Thermo complete.

+++ main/model/engine.h
@@ -32,7 +32,7 @@ typedef struct {
    * @brief The engine's external temperature in Celsius degrees.
    *
    */
-  double temp_c;
+  float temp_c;
   bool temp_c_available;
 #endif
